properly split up boolean_ops tests
